Removed a bug. See extended description for more info.
The bug was created from the problem of not checking if ImageView was a
null refferenced object. If you haven't chosen a picture the ImageView
WPF form is never called and calling it's called method is throwing a
NullRefferenceException. Now you just check the object and if it's not
null, you close it.